+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix remove wallet dialog button padding
 - Add cancel button to disable pin access dialog
 - Disable pin input when restoring a wallet
+- Disable firmware update dialog next button after being clicked
 
 ## [1.1.3] - 2022-07-18
 ### Added